Addding a separate NBF file for BP
